+++ src/http/hpack_huffman.c
@@ -131,7 +131,7 @@ NODISCARD static HuffmanDecodeResult decode_bytes_huffman_impl(const HuffmanTree
 		    TSTR_STATIC_LIT("8 or more bits not decoded, is also invalid"));
 	}
 
-	uint8_t bits_mask = ((1 << bits_not_decoded) - 1);
+	uint8_t bits_mask = (uint8_t)((1U << bits_not_decoded) - 1U);
 
 	if((data[last_pos.pos] & bits_mask) != (EOS_BYTE & bits_mask)) {
 		// last not decoded bits are not the eos bytes
@@ -253,9 +253,9 @@ hpack_huffman_encode_value_fixed_size_impl(const HuffmanEncodeMap* const map, vo
 			const uint8_t bit = value & 0x01;
 			value = value >> 1;
 
-			data_ptr[current_pos.pos] =
-			    data_ptr[current_pos.pos] | (bit << (7 - // NOLINT(readability-magic-numbers)
-			                                         current_pos.bits_pos));
+			data_ptr[current_pos.pos] = data_ptr[current_pos.pos] |
+			                            (uint8_t)(bit << (7U - // NOLINT(readability-magic-numbers)
+			                                              current_pos.bits_pos));
 
 			bit_pos_inc(&current_pos);
 		}
@@ -266,7 +266,8 @@ hpack_huffman_encode_value_fixed_size_impl(const HuffmanEncodeMap* const map, vo
 	if(current_pos.bits_pos > 0) {
 		const size_t missing_bits = 8 - current_pos.bits_pos;
 		// set the last bits to EOS bits (all 1s)
-		data_ptr[current_pos.pos] = data_ptr[current_pos.pos] | ((1 << missing_bits) - 1);
+		data_ptr[current_pos.pos] =
+		    data_ptr[current_pos.pos] | (uint8_t)((1U << missing_bits) - 1U);
 
 		current_pos.bits_pos = 0;
 		current_pos.pos += 1;
